+++ src/main/java/woowacourse/shoppingcart/application/ProductService.java
@@ -6,8 +6,8 @@ import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Product;
 
 import java.util.List;
-import woowacourse.shoppingcart.dto.ProductRequest;
-import woowacourse.shoppingcart.dto.ProductResponse;
+import woowacourse.shoppingcart.dto.request.ProductRequest;
+import woowacourse.shoppingcart.dto.response.ProductResponse;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
