+++ src/main/java/woowacourse/shoppingcart/domain/Orders.java
@@ -7,7 +7,8 @@ public class Orders {
     private Long id;
     private List<OrderDetail> orderDetails;
 
-    public Orders() {}
+    public Orders() {
+    }
 
     public Orders(final Long id, final List<OrderDetail> orderDetails) {
         this.id = id;
