+++ src/main/java/woowacourse/shoppingcart/dto/request/CustomerRequest.java
@@ -1,22 +1,28 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.request;
 
 import woowacourse.shoppingcart.domain.Customer;
 
-public class CustomerUpdateRequest {
+public class CustomerRequest {
 
+    private String loginId;
     private String name;
     private String password;
 
-    public CustomerUpdateRequest() {
+    public CustomerRequest() {
     }
 
-    public CustomerUpdateRequest(String name, String password) {
+    public CustomerRequest(String loginId, String name, String password) {
+        this.loginId = loginId;
         this.name = name;
         this.password = password;
     }
 
-    public Customer toCustomer(String loginId) {
+    public Customer toCustomer() {
         return new Customer(loginId, name, password);
+    }
+
+    public String getLoginId() {
+        return loginId;
     }
 
     public String getName() {
