+++ src/main/java/woowacourse/shoppingcart/dto/response/OrderDetailsResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
